Remove old unused files.
