+++ src/lib/mino/data.ts
@@ -1,193 +1,107 @@
-/**
- * Functions related to the direct manipulation of the underlying polyomino data.
- *
- * Minos are encoded as a combination of a value and a width flag:
- *
- * 0100_0111_0100
- * ^ value   ^ width
- *
- * This describes the L tetromino:
- *
- * []
- * [][][]
- *
- * The last four bits represent the width of the mino and how the remaining bits
- * should be interpreted. The value `11_0001` represents the vertical domino:
- *
- * []
- * []
- *
- * while the value `11_0010` represents the horizontal domino:
- *
- * [][]
- *
- * This encoding was chosen so that operations on polyominoes can be done using efficient
- * bit-shift operations instead of heavier operations on arrays and sets.
- */
+import { range } from "lodash-es"
+import Vector, { type VectorLike } from "../vector"
+import type Polyomino from "./Polyomino"
 
-import Vector, { type VectorLike } from "$lib/vector"
-
-// type for the encoded representation of a mino
-export type MinoData = number
-// type for the coordinates of a mino square
-export type Coord = Vector
-// type for the dimensions of a mino
 export type Dims = [number, number]
+export type Coord = Vector
 
-// The number of bits reserved for the width
-export const WIDTH_BITS = 4
-export const MAX_WIDTH = 1 << WIDTH_BITS
-
-/**
- * Return the raw data portion of the mino.
- */
-export function getData(mino: MinoData): number {
-  return mino >> WIDTH_BITS
+export interface PossibleRelativeLink {
+  mino?: Polyomino
+  coord: Vector
 }
 
-export function getOrder(mino: MinoData): number {
-  let data = getData(mino)
-  let size = 0
-  while (data) {
-    size += data & 1
-    data = data >> 1
-  }
-  return size
+export type RelativeLink = Required<PossibleRelativeLink>
+
+export type MinoData = Set<number>
+
+export function mask(x: number, y: number) {
+  return (x << 16) | y
+}
+export function maskVec([x, y]: VectorLike) {
+  return mask(x, y)
+}
+
+export function mx(mask: number) {
+  return mask >> 16
+}
+export function my(mask: number) {
+  return mask % (1 << 16)
+}
+
+export function unmask(m: number): [number, number] {
+  return [mx(m), my(m)]
 }
 
 export function getWidth(mino: MinoData) {
-  return mino % MAX_WIDTH || MAX_WIDTH
+  let min = Infinity
+  let max = -Infinity
+  for (let value of mino.values()) {
+    min = Math.min(min, mx(value))
+    max = Math.max(max, mx(value))
+  }
+  return max - min + 1
 }
 
 export function getHeight(mino: MinoData) {
-  const w = getWidth(mino)
-  const data = getData(mino)
-  return Math.floor(Math.log2(data) / w) + 1
+  let min = Infinity
+  let max = -Infinity
+  for (let value of mino.values()) {
+    min = Math.min(min, my(value))
+    max = Math.max(max, my(value))
+  }
+  return max - min + 1
 }
 
-/**
- * Return the mino given the data bits and the specified width
- */
-export function fromBits(data: number, width: number): MinoData {
-  return (data << WIDTH_BITS) | (width === MAX_WIDTH ? 0 : width)
+export function getKey(data: MinoData) {
+  const xs = [...data.values()].sort()
+  return xs.join(",")
 }
 
-/**
- * Iterate over the coordinates of the squares of the mino.
- */
-export function* getCoords(mino: MinoData): Generator<Coord> {
-  let data = getData(mino)
-  const w = getWidth(mino)
-  let k = 0
-  while (data) {
-    if (data & 1) {
-      yield new Vector((k / w) >> 0, k % w)
-    }
-    k++
-    data = data >> 1
+export function addAll(mino: MinoData, items: Iterable<VectorLike>) {
+  for (const p of items) {
+    mino.add(maskVec(p))
   }
 }
 
-/**
- * Create a mino given a list of coordinates.
- */
-export function fromCoords(coords: VectorLike[]) {
-  const w = Math.max(...coords.map(([, y]) => y)) + 1
-  let result = 0
-  for (const [x, y] of coords) {
-    result = result | (1 << (w * x + y))
-  }
-  return fromBits(result, w)
+export function display(
+  mino: MinoData,
+  entry = "1",
+  empty = "0",
+  delimiter = "_",
+) {
+  return range(getHeight(mino))
+    .map((y) => {
+      return range(getWidth(mino))
+        .map((x) => (mino.has(mask(x, y)) ? entry : empty))
+        .join("")
+    })
+    .join(delimiter)
+}
+
+export function hasX(mino: MinoData, x: number) {
+  return mino.values().some((v) => mx(v) === x)
+}
+
+export function hasY(mino: MinoData, y: number) {
+  return mino.values().some((v) => my(v) === y)
 }
 
 /**
- * Get the coord `[i, j]` in the mino with width `w`.
+ * Return the neighbors of the coord [i,j]
  */
-export function getCoordMask(i: number, j: number, w: number) {
-  return 1 << (i * w + j + WIDTH_BITS)
+export function* getNeighbors(p: Vector): Generator<Vector> {
+  // TODO it turns out this order greatly impacts the order of the minos
+  // either standardize it or sort the minos independently
+  yield p.add(Vector.DOWN)
+  yield p.add(Vector.UP)
+  yield p.add(Vector.RIGHT)
+  yield p.add(Vector.LEFT)
 }
 
-/**
- * Returns true if the mino contains the given coordinate
- */
-export function contains(mino: MinoData, [x, y]: VectorLike) {
-  const w = getWidth(mino)
-  if (x < 0 || y < 0 || y >= w) {
-    return false
-  }
-  return !!(mino & getCoordMask(x, y, w))
-}
-
-/**
- * Get the bitmask corresponding to the jth column of the mino
- */
-export function getColumnMask(mino: MinoData, j: number): number {
-  const width = getWidth(mino)
-  const height = getHeight(mino)
-  let mask = 0
-  for (let i = 0; i < height; i++) {
-    mask |= getCoordMask(i, j, width)
-  }
-  return mino & mask
-}
-
-/**
- * Get the rows of the mino from bottom-up
- */
-export function* rowBits(mino: MinoData): Generator<number> {
-  const w = getWidth(mino)
-  let data = getData(mino)
-  while (data) {
-    yield data % (1 << w)
-    data >>= w
-  }
-}
-
-/**
- * Return the mino represented by the given delimited string:
- * e.g.
- * 100_111 =>
- * []
- * [][][]
- */
-export function fromString(str: string) {
-  const width = str.split("_")[0].length
-  const bits = parseInt(str.replace(/_/g, ""), 2)
-  return fromBits(bits, width)
-}
-
-interface DisplayOpts {
-  block?: string
-  space?: string
-}
-
-/**
- * Get the string representation of a mino,
- * using the same delimited expression as `fromString()`.
- */
-export function toString(mino: MinoData) {
-  const w = getWidth(mino)
-  const strs: string[] = []
-  for (const row of rowBits(mino)) {
-    strs.push(row.toString(2).padStart(w, "0"))
-  }
-  return strs.reverse().join("_")
-}
-
-/**
- * Return a pretty printed representation of the polyomino
- *
- * @param mino the polyomino object
- * @param block the string used to represent a filled block
- * @param space the string used to represent an unfilled block
- */
-export function displayMino(mino: MinoData, opts: DisplayOpts = {}) {
-  const { block = "□", space = " " } = opts
-  const w = getWidth(mino)
-  const result = []
-  for (const row of rowBits(mino)) {
-    const str = row.toString(2).padStart(w, "0")
-    result.push([...str].join(" "))
-  }
-  return result.reverse().join("\n").replace(/1/g, block).replace(/0/g, space)
+export function* getKingwiseNeighbors(p: Vector): Generator<Vector> {
+  yield* getNeighbors(p)
+  yield p.add([1, 1])
+  yield p.add([1, -1])
+  yield p.add([-1, -1])
+  yield p.add([-1, 1])
 }
